Fix `extraCommands` and `extraStopCommands`
`networking.firewall.extraCommands` and `extraStopCommands` properties
were used to provide as a way to intoduce custom logic into the firewall
via a shell script. This commit keeps `extraCommands` and
`extraStopCommands` as they are and provides an alternative way to
customize the firewall behaviour via the `extraRules` and the
`preRefuseRules` property.

The shell commands within `extraCommands` and `extraStopCommands` will
still get run as part of the script that starts/stops/reloads the
firewall service. As an additional safety measure, if any of the
afformentioned properties contain "iptables" the module will generate a
warning informing that the firewall *may* not work as intended.

The `extraRules` option takes over as the nftables-specific way to add
custom rules to the firewall.

The rules added via `preRefuseRules` lives in the `nixos-pre-refuse` chain
of the filter table and runs any user provided rules on the packets that
are about to be refused. Therefore, the newly added `nixos-pre-refuse`
chain will contains all the user-provided rules followed by a `jump` to
`nixos-fw-log-refuse`.