Update the service to remove rules instead of flushing ruleset
Flushing the entire ruleset means that anything that any change to
nftables has to be done by interacting with the `networking.firewall`
module. This means that things like natting, tproxying or any nft-based
configuration that you apply will be tightly coupled to the firewall and
cannot be enabled/disabled independently from the firewall.