Create base chains on our own
The module previously relied on a helper function to generate
iptables-like base tables which are 'entry points' and where nft starts
applying rules for packets.

A big disdvatage using that method was that removing rules were always
awkward as nftables only supprots removing rules via handles. Handles
are assigned to tables, chains, rules and they cannot be modified. So to
remove a particular rule from the input chain for example, one would
have to grab a handle for the rule they would like to remove which
requires a either text processing via the shell, or (thankfully) json
processing via something like jq. However, adding jq as a dependency to
the firewall just didn't feel right.

The obvious solution was to define base chains by ourselves in a
namespaced table. Using a namespaced table means that cleanup becomes a
lot easier. All you have to do is flush and delete the table. It makes
the final ruleset a lot more concise as well. The only issue with this
would be that there is no defined order to execute base rules if
multiple base rules have the same priority values. Therefore, a future
change will allow users to define a `priorityOffset` which can offset
the rule priority using a signed numeric value.