Fix NAT module
The firewall module present in nixpkgs is tightly coupled with the nat
module. If enabled alone, the NAT module will not create a separate
module and instead, just piggyback on the firewall module's properties
to add arbitrary shell commands to set itself up.
Since we're using nftables, such a dependency is not required.

This change will properly set up the systemd service and clean up the
NAT rules a bit.